22May24_1919L
I worked on the data builder for the MERLIN dataset, to get it gridded for the 64x64 HRRR grid. Some considerations include to have a 65x65 HRRR grid fed into the midpoint and binning calculations so I do not have to pad the outside with zeros.  Also, maybe have a 64x64 zoomed out HRRR 500mb/700mb flow and temperature panels to capture the vertical nature of the sea breaze.  Additionally, it will be important to sort the sources and run them through the flash sorting algorithm before building the lightning data structure.