23May24_1749L
This is close to being finished.  I need to plot a scatter plot diagram of the midpoints and the actual HRRR grids. This script also writes the data to my scratch directory, and uses a 65x65 HRRR grid, with one row added from the south, and east, to build the gridded lightning datasets.  Currently this script only runs for the month of June 2022.  More data needs to be collected.